breaking up TwitchRepoImpl
